Add update for profil and security route.

+++ src/main/java/com/epitech/pictmanager/components/user/services/UserService.java
@@ -8,10 +8,14 @@ import com.epitech.pictmanager.components.user.repositories.UserJpaRepository;
 import com.epitech.pictmanager.models.Profil;
 import com.epitech.pictmanager.models.User;
 import com.epitech.pictmanager.responses.GenericResponse;
+import com.epitech.pictmanager.responses.GenericUpdateResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
+import java.util.List;
 
 @Service
 public class UserService {
@@ -29,9 +33,12 @@ public class UserService {
     public ResponseEntity<GenericResponse> deleteUserAndProfil(String id) {
         try {
             User user = this.userRepository.findUserById(Long.parseLong(id));
-            Profil profile = this.profileRepository.findProfileByUser(user);
             if (user == null) {
                 throw new RuntimeException("User not found");
+            }
+            Profil profile = this.profileRepository.findProfileByUser(user);
+            if(profile == null) {
+                throw new RuntimeException("Profile not found");
             }
             this.profileRepository.delete(profile);
             this.userRepository.delete(user);
@@ -41,9 +48,9 @@ public class UserService {
         }
     }
 
-    public ResponseEntity<GenericResponse> updateUserProfil(String username, UpdateProfilDto updateProfilDto) {
+    public ResponseEntity<GenericUpdateResponse> updateUserProfil(String id, UpdateProfilDto updateProfilDto) {
         try {
-            User user = this.userRepository.findUserByUsername(username);
+            User user = this.userRepository.findUserById(Long.parseLong(id));
             if (user == null) {
                 throw new RuntimeException("User not found");
             }
@@ -51,38 +58,73 @@ public class UserService {
             if (profil == null) {
                 throw new RuntimeException("Profile not found");
             }
+            List<String> updatedFields = new ArrayList<>();
             boolean hasChanged = false;
             if (updateProfilDto.getDescription() != null && !updateProfilDto.getDescription().equals(profil.getDescription())) {
                 profil.setDescription(updateProfilDto.getDescription());
+                updatedFields.add("description");
                 hasChanged = true;
             }
             if (updateProfilDto.getProfilePicture() != null && !updateProfilDto.getProfilePicture().equals(profil.getProfilePicture())) {
                 profil.setProfilePicture(updateProfilDto.getProfilePicture());
+                updatedFields.add("profilePicture");
                 hasChanged = true;
             }
             if (updateProfilDto.getCoverPicture() != null && !updateProfilDto.getCoverPicture().equals(profil.getCoverPicture())) {
                 profil.setCoverPicture(updateProfilDto.getCoverPicture());
+                updatedFields.add("coverPicture");
                 hasChanged = true;
             }
             if (!hasChanged) {
-                return new ResponseEntity<GenericResponse>(new GenericResponse("No changes detected, profile not updated", HttpStatus.NOT_MODIFIED.value()), HttpStatus.NOT_MODIFIED);
+                return new ResponseEntity<GenericUpdateResponse>(new GenericUpdateResponse("No changes detected, profile not updated", HttpStatus.NOT_MODIFIED.value(), updatedFields), HttpStatus.NOT_MODIFIED);
             }
             this.profileRepository.save(profil);
-            return new ResponseEntity<GenericResponse>(new GenericResponse("Profile updated successfully", HttpStatus.ACCEPTED.value()), HttpStatus.ACCEPTED);
+            return new ResponseEntity<GenericUpdateResponse>(new GenericUpdateResponse("Profile updated successfully", HttpStatus.ACCEPTED.value(), updatedFields), HttpStatus.ACCEPTED);
         } catch (RuntimeException e) {
             throw new RuntimeException(e.getMessage());
         }
     }
 
-    public ResponseEntity<GenericResponse> updateUserSecurity(String username, UpdateSecurityDto updateSecurityDto) {
+    public ResponseEntity<GenericUpdateResponse> updateUserSecurity(String id, UpdateSecurityDto updateSecurityDto) {
         try {
-            User user = this.userRepository.findUserByUsername(username);
+            User user = this.userRepository.findUserById(Long.parseLong(id));
+            boolean isPublic = updateSecurityDto.getVisibility() == "true" ? true : false;
             if (user == null) {
                 throw new RuntimeException("User not found");
             }
             boolean hasChanged = false;
-
-            return new ResponseEntity<GenericResponse>(new GenericResponse("Profile updated successfully", HttpStatus.ACCEPTED.value()), HttpStatus.ACCEPTED);
+            List<String> updatedFields = new ArrayList<>();
+            if (!user.getUsername().equals(updateSecurityDto.getUsername())) {
+                if(this.userRepository.existsUserByUsername(updateSecurityDto.getUsername())) {
+                    throw new RuntimeException("Username already taken");
+                }
+                user.setUsername(updateSecurityDto.getUsername());
+                updatedFields.add("username");
+                hasChanged = true;
+            }
+            if (!user.getEmail().equals(updateSecurityDto.getEmail())) {
+                if(this.userRepository.existsUserByEmail(updateSecurityDto.getEmail())) {
+                    throw new RuntimeException("Email already taken");
+                }
+                user.setEmail(updateSecurityDto.getEmail());
+                updatedFields.add("email");
+                hasChanged = true;
+            }
+            if (!user.getPublic().equals(isPublic)) {
+                user.setPublic(isPublic);
+                updatedFields.add("isPublic");
+                hasChanged = true;
+            }
+            if(!passwordEncryptionService.check(updateSecurityDto.getPassword(), user.getPassword())) {
+                user.setPassword(passwordEncryptionService.encrypt(updateSecurityDto.getPassword()));
+                updatedFields.add("password");
+                hasChanged = true;
+            }
+            if(!hasChanged) {
+                return new ResponseEntity<GenericUpdateResponse>(new GenericUpdateResponse("No changes detected, profile not updated", HttpStatus.NOT_MODIFIED.value(), updatedFields), HttpStatus.NOT_MODIFIED);
+            }
+            this.userRepository.save(user);
+            return new ResponseEntity<GenericUpdateResponse>(new GenericUpdateResponse("Profile updated successfully", HttpStatus.ACCEPTED.value(), updatedFields), HttpStatus.ACCEPTED);
         } catch (RuntimeException e) {
             throw new RuntimeException(e.getMessage());
         }
